chore(workspace): add publish scripts and docs
