+++ .commitlintrc.json
@@ -1,4 +1,4 @@
 {
-  "extends": ["@commitlint/config-angular"],
+  "extends": ["@commitlint/config-conventional"],
   "rules": {}
 }
