+++ .verdaccio/config.yml
@@ -19,7 +19,7 @@ packages:
     proxy: npmjs
 
 # log settings
-logs:
+log:
   type: stdout
   format: pretty
   level: warn
