+++ presto-client/project.json
@@ -6,37 +6,68 @@
   "targets": {
     "build": {
       "executor": "@nx/js:swc",
-      "outputs": ["{options.outputPath}"],
       "options": {
-        "outputPath": "dist/presto-client",
+        "assets": ["presto-client/*.md"],
         "main": "presto-client/src/index.ts",
-        "tsConfig": "presto-client/tsconfig.lib.json",
-        "assets": ["presto-client/*.md"]
+        "outputPath": "dist/presto-client",
+        "tsConfig": "presto-client/tsconfig.lib.json"
+      },
+      "outputs": ["{options.outputPath}"]
+    },
+    "github": {
+      "executor": "@jscutlery/semver:github",
+      "options": {
+        "draft": true,
+        "generateNotes": true,
+        "notes": "${notes}",
+        "tag": "${tag}"
       }
     },
     "lint": {
       "executor": "@nx/linter:eslint",
-      "outputs": ["{options.outputFile}"],
       "options": {
         "lintFilePatterns": ["presto-client/**/*.ts", "presto-client/package.json"]
-      }
+      },
+      "outputs": ["{options.outputFile}"]
     },
-    "publish": {
-      "command": "node tools/scripts/publish.mjs presto-client {args.ver} {args.tag}",
-      "dependsOn": ["build", "lint", "test"]
+    "npm": {
+      "command": "node tools/scripts/npm.mjs presto-client ${version} ${tag}"
     },
     "test": {
-      "executor": "@nx/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"],
-      "options": {
-        "jestConfig": "presto-client/jest.config.ts",
-        "passWithNoTests": true
-      },
       "configurations": {
         "ci": {
           "ci": true,
           "codeCoverage": true
         }
+      },
+      "executor": "@nx/jest:jest",
+      "options": {
+        "jestConfig": "presto-client/jest.config.ts",
+        "passWithNoTests": true
+      },
+      "outputs": ["{workspaceRoot}/coverage/{projectRoot}"]
+    },
+    "version": {
+      "dependsOn": ["build", "lint", "test"],
+      "executor": "@jscutlery/semver:version",
+      "options": {
+        "commitMessageFormat": "chore({projectName}): release ${version}",
+        "postTargets": ["presto-client:github", "presto-client:npm"],
+        "preset": "conventional",
+        "push": true,
+        "skipCommitTypes": ["build", "ci", "chore", "docs", "style", "test"]
+      }
+    },
+    "version-local": {
+      "dependsOn": ["build", "lint", "test"],
+      "executor": "@jscutlery/semver:version",
+      "options": {
+        "commitMessageFormat": "chore({projectName}): release ${version}",
+        "postTargets": ["presto-client:npm"],
+        "preset": "conventional",
+        "push": false,
+        "skipCommit": true,
+        "skipCommitTypes": ["build", "ci", "chore", "docs", "style", "test"]
       }
     }
   },
