+++ presto-client/tsconfig.json
@@ -1,12 +1,13 @@
 {
   "compilerOptions": {
     "forceConsistentCasingInFileNames": true,
-    "module": "esnext",
+    "module": "CommonJS",
+    "noFallthroughCasesInSwitch": true,
     "noImplicitOverride": true,
     "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true,
     "noPropertyAccessFromIndexSignature": true,
-    "strict": true
+    "strict": true,
+    "target": "ESNext"
   },
   "extends": "../tsconfig.base.json",
   "files": [],
