+++ project.json
@@ -5,8 +5,8 @@
     "local-registry": {
       "executor": "@nx/js:verdaccio",
       "options": {
-        "port": 4873,
         "config": ".verdaccio/config.yml",
+        "port": 4873,
         "storage": "tmp/local-registry/storage"
       }
     }
